// frontend/src/types/cv.ts
export interface ProjectItem {
  name: string;
  description: string;
  tech_stack: string[];
  repo_url?: string;
}
export interface ProjectsContent {
  items: ProjectItem[];
}